Crate burn_fusion

Burn Fusion

This library is a part of the Burn project. It is a standalone crate that can be used to perform automatic operation fusion on backends that support it.

Modules

client

Client module exposing types to communicate with the fusion server.

stream

Stream module exposing all tensor operations that can be optimized.

Macros

binary_float_cmp_ops

binary_float_ops

binary_int_cmp_ops

binary_int_ops

scalar_float2int_ops

scalar_float_cmp_ops

scalar_float_ops

scalar_int_cmp_ops

scalar_int_ops

unary_float_ops

unary_int_ops

Structs

Fusion

Enable dynamic operation fusion on a backend that implements fusion backend.

FusionTensor

Tensor primitive for the fusion backend for all kind.

OptimizationProperties

The properties of a builder.

Enums

OptimizationStatus

The status of a builder.

Traits

FusionBackend

Trait that allows an existing backend to specify graph optimizations using operation builder.

FusionRuntime

Trait that defines a runtime that will benefits from fused operations.

Optimization

The operation created from the builder.

OptimizationBuilder

The fusion operation abstraction allows implementations to fuse many tensor operations into one, improving the performance of the backend.

Type Aliases

Client

Type alias for <R as FusionRuntime>::FusionClient.

FusionDevice

Type alias for <R as FusionRuntime>::FusionDevice.

FusionHandle

Type alias for <R as FusionRuntime>::FusionHandle.

Handle

Type alias for representation backend handle.