Module burn_fusion::graph

Graph module exposing all tensor operations that can be optimized.

Structs

AdaptiveAvgPool1dBackwardDescription

AdaptiveAvgPool1dDescription

AdaptiveAvgPool2dBackwardDescription

AdaptiveAvgPool2dDescription

AvgPool1dBackwardDescription

AvgPool1dDescription

AvgPool2dBackwardDescription

AvgPool2dDescription

BinaryOpsDescription

CatOpsDescription

ClampOpsDescription

Context

The context contains the relative graph tensor mapping so that a relative tensor id can be mapped to an existing tensor that can be fetched and updated with the handle container.

Conv1dDescription

Conv2dDescription

ConvTranspose1dDescription

ConvTranspose2dDescription

EmbeddingBackwardDescription

EmbeddingDescription

GatherOpsDescription

Graph

The computational graph containing a list of tensor operation descriptions.

MaskFillOpsDescription

MaskWhereOpsDescription

MaxPool1dDescription

MaxPool1dWithIndicesBackwardDescription

MaxPool1dWithIndicesDescription

MaxPool2dDescription

MaxPool2dWithIndicesBackwardDescription

MaxPool2dWithIndicesDescription

RandomOpsDescription

ReduceDimWithIndicesDescription

RepeatOpsDescription

ReshapeDescription

ScalarOpsDescription

ScatterOpsDescription

SelectAssignOpsDescription

SelectOpsDescription

SliceAssignOpsDescription

SliceOpsDescription

SwapDimsDescription

Swap dim operation description.

UnaryOpsDescription

Enums

BaseOpsDescription

Basic operations that can be done on any tensor type.

BoolOpsDescription

Operation description specific to a bool tensor.

CacheResult

Action to be made depending on the graph.

Condition

When checking if an optimization is possible, a start or an end condition ensures that this optimization is always optimal.

FloatOpsDescription

Operation description specific to a float tensor.

IntOpsDescription

Operation description specific to an int tensor.

ModuleOpsDescription

Operation description specific to module.

NumericOpsDescription

Numeric operations on int and float tensors.

TensorOpsDescription

Describe all tensor operations possible.

Traits

Ops

General trait to abstract how a single operation is executed.