Module burn_fusion::stream

Stream module exposing all tensor operations that can be optimized.

Structs

AdaptiveAvgPool1dBackwardDescription

AdaptiveAvgPool1dDescription

AdaptiveAvgPool2dBackwardDescription

AdaptiveAvgPool2dDescription

AvgPool1dBackwardDescription

AvgPool1dDescription

AvgPool2dBackwardDescription

AvgPool2dDescription

BinaryOperationDescription

CatOperationDescription

ClampOperationDescription

Context

The context contains the relative graph tensor mapping so that a relative tensor id can be mapped to an existing tensor that can be fetched and updated with the handle container.

Conv1dDescription

Conv1dOptionsDescription

Conv2dDescription

Conv2dOptionsDescription

ConvTranspose1dDescription

ConvTranspose1dOptionsDescription

ConvTranspose2dDescription

ConvTranspose2dOptionsDescription

EmbeddingBackwardDescription

EmbeddingDescription

GatherOperationDescription

MaskFillOperationDescription

MaskWhereOperationDescription

MaxPool1dDescription

MaxPool1dWithIndicesBackwardDescription

MaxPool1dWithIndicesDescription

MaxPool2dDescription

MaxPool2dWithIndicesBackwardDescription

MaxPool2dWithIndicesDescription

MultiStream

Keep track of multiple concurrent streams of operations.

OperationQueue

A growing list of tensor operation descriptions.

RandomOperationDescription

ReduceDimWithIndicesDescription

RepeatOperationDescription

ReshapeDescription

ScalarOperationDescription

ScatterOperationDescription

SelectAssignOperationDescription

SelectOperationDescription

SliceAssignOperationDescription

SliceOperationDescription

StreamId

The stream id.

SwapDimsDescription

Swap dim operation description.

UnaryOperationDescription

Enums

BaseOperationDescription

Basic operations that can be done on any tensor type.

BoolOperationDescription

Operation description specific to a bool tensor.

FloatOperationDescription

Operation description specific to a float tensor.

IntOperationDescription

Operation description specific to an int tensor.

ModuleOperationDescription

Operation description specific to module.

NumericOperationDescription

Numeric operations on int and float tensors.

OperationDescription

Describe all tensor operations possible.

Traits

Operation

General trait to abstract how a single operation is executed.