burn_fusion::stream

Struct Context

pub struct Context<'a, H> {

    pub tensors: &'a mut HashMap<TensorId, TensorDescription>,
    pub handles: &'a mut HandleContainer<H>,
    pub scalar_f32: &'a Vec<f32>,
    pub scalar_f16: &'a Vec<f16>,
    pub scalar_bf16: &'a Vec<bf16>,
    pub scalar_i64: &'a Vec<i64>,
    pub scalar_i32: &'a Vec<i32>,
    pub scalar_i16: &'a Vec<i16>,
    pub scalar_i8: &'a Vec<i8>,
    pub scalar_u64: &'a Vec<u64>,
    pub scalar_u32: &'a Vec<u32>,
    pub scalar_u16: &'a Vec<u16>,
    pub scalar_u8: &'a Vec<u8>,

}

The context contains the relative graph tensor mapping so that a relative tensor id can be mapped to an existing tensor that can be fetched and updated with the handle container.

It also contains all scalar values, which can change even for the same graph. They are sorted in the order in which they appear in the graph.

Fields
tensors: &'a mut HashMap<TensorId, TensorDescription>

The tensor mapping where local tensor id points to the updated tensor description.

handles: &'a mut HandleContainer<H>

Handle container to retrieve tensors based on their description.

scalar_f32: &'a Vec<f32>

F32 scalars found in the graph in the order they appeared.

scalar_f16: &'a Vec<f16>

F16 scalars found in the graph in the order they appeared.

scalar_bf16: &'a Vec<bf16>

BF16 scalars found in the graph in the order they appeared.

scalar_i64: &'a Vec<i64>

i64 scalars found in the graph in the order they appeared.

scalar_i32: &'a Vec<i32>

i32 scalars found in the graph in the order they appeared.

scalar_i16: &'a Vec<i16>

i16 scalars found in the graph in the order they appeared.

scalar_i8: &'a Vec<i8>

i8 scalars found in the graph in the order they appeared.

scalar_u64: &'a Vec<u64>

u64 scalars found in the graph in the order they appeared.

scalar_u32: &'a Vec<u32>

u32 scalars found in the graph in the order they appeared.

scalar_u16: &'a Vec<u16>

u16 scalars found in the graph in the order they appeared.

scalar_u8: &'a Vec<u8>

u8 scalars found in the graph in the order they appeared.

Implementations

impl<'a, H> Context<'a, H>

pub fn new( tensors: &'a mut HashMap<TensorId, TensorDescription>, handles: &'a mut HandleContainer<H>, scalar_f32: &'a Vec<f32>, scalar_f16: &'a Vec<f16>, scalar_bf16: &'a Vec<bf16>, scalar_i64: &'a Vec<i64>, scalar_i32: &'a Vec<i32>, scalar_i16: &'a Vec<i16>, scalar_i8: &'a Vec<i8>, scalar_u64: &'a Vec<u64>, scalar_u32: &'a Vec<u32>, scalar_u16: &'a Vec<u16>, scalar_u8: &'a Vec<u8>, ) -> Self

Constructs a new Context.