Trait burn_fusion::FusionBackend

pub trait FusionBackend: Backend {
    type FusionDevice: FusionDevice + From<Self::Device> + Into<Self::Device> + Debug;
    type Handle: Sync + Send + Clone;
    type FusionClient: FusionClient<FusionBackend = Self>;

    // Required methods
    fn optimizations(
        device: &Device<Self>
    ) -> Vec<Box<dyn OptimizationBuilder<Self>>>;
    fn float_tensor<const D: usize>(
        handle: Self::Handle,
        shape: Shape<D>
    ) -> Self::TensorPrimitive<D>;
    fn int_tensor<const D: usize>(
        handle: Self::Handle,
        shape: Shape<D>
    ) -> Self::IntTensorPrimitive<D>;
    fn bool_tensor<const D: usize>(
        handle: Self::Handle,
        shape: Shape<D>
    ) -> Self::BoolTensorPrimitive<D>;
    fn float_tensor_handle<const D: usize>(
        tensor: Self::TensorPrimitive<D>
    ) -> Self::Handle;
    fn int_tensor_handle<const D: usize>(
        tensor: Self::IntTensorPrimitive<D>
    ) -> Self::Handle;
    fn bool_tensor_handle<const D: usize>(
        tensor: Self::BoolTensorPrimitive<D>
    ) -> Self::Handle;
}

Trait that allows an existing backend to specify graph optimizations using operation builder.

Required Associated Types

type FusionDevice: FusionDevice + From<Self::Device> + Into<Self::Device> + Debug

The device type that can return an ID.

It can be the same as (Backend::Device), but must implement (FusionDevice).

type Handle: Sync + Send + Clone

The type that can be used to point to a tensor of any kind.

type FusionClient: FusionClient<FusionBackend = Self>

What kind of client should be used.

Required Methods

fn optimizations( device: &Device<Self> ) -> Vec<Box<dyn OptimizationBuilder<Self>>>

The list of optimizations that will be used to optimize the computational graph.

fn float_tensor<const D: usize>( handle: Self::Handle, shape: Shape<D> ) -> Self::TensorPrimitive<D>

Convert a handle to a float tensor.

fn int_tensor<const D: usize>( handle: Self::Handle, shape: Shape<D> ) -> Self::IntTensorPrimitive<D>

Convert a handle to an int tensor.

fn bool_tensor<const D: usize>( handle: Self::Handle, shape: Shape<D> ) -> Self::BoolTensorPrimitive<D>

Convert a handle to a bool tensor.

fn float_tensor_handle<const D: usize>( tensor: Self::TensorPrimitive<D> ) -> Self::Handle

Convert a float tensor to a handle.

fn int_tensor_handle<const D: usize>( tensor: Self::IntTensorPrimitive<D> ) -> Self::Handle

Convert an int tensor to a handle.

fn bool_tensor_handle<const D: usize>( tensor: Self::BoolTensorPrimitive<D> ) -> Self::Handle

Convert a bool tensor to a handle.