burn_fusion

Trait FusionRuntime

pub trait FusionRuntime:
    Send
    + Sync
    + Sized
    + Debug {
    type OptimizationState: Serialize + DeserializeOwned;
    type Optimization: Optimization<Self>;
    type FusionHandle: Clone + Send;
    type FusionDevice: DeviceOps;
    type FusionClient: FusionClient<Self>;
    type BoolRepr: Element;

    // Required method
    fn optimizations(
        device: Self::FusionDevice,
    ) -> Vec<Box<dyn OptimizationBuilder<Self::Optimization>>>;
}

Trait that defines a runtime that will benefits from fused operations.

Required Associated Types

type OptimizationState: Serialize + DeserializeOwned

The state that can be serialized for an optimization.

type Optimization: Optimization<Self>

Optimization type for the backend.

type FusionHandle: Clone + Send

Handle used to store tensor dynamically.

type FusionDevice: DeviceOps

Device used by the runtime.

type FusionClient: FusionClient<Self>

The client to interact with the runtime.

type BoolRepr: Element

The type that represents booleans on the backend.

Required Methods

fn optimizations( device: Self::FusionDevice, ) -> Vec<Box<dyn OptimizationBuilder<Self::Optimization>>>

The list of optimizations that will be used to optimize the computational graph.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.