burn_fusion

Trait OptimizationBuilder

pub trait OptimizationBuilder<O>: Send {
    // Required methods
    fn register(&mut self, operation: &OperationDescription);
    fn build(&self) -> O;
    fn reset(&mut self);
    fn status(&self) -> OptimizationStatus;
    fn properties(&self) -> OptimizationProperties;
    fn len(&self) -> usize;

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

The fusion operation abstraction allows implementations to fuse many tensor operations into one, improving the performance of the backend.

Notes

The implementations are free to execute the registered operations the way they want to improve the speed and efficiency of the computational graph. It doesn’t mean that all registered operations should be fused, but that another way of executing them is more efficient.

Also, it is important to return (OptimizationStatus::Closed) when no more registered operation can improve the performance.

Required Methods

fn register(&mut self, operation: &OperationDescription)

Register a new tensor operation.

fn build(&self) -> O

Finish the optimization and create a fusion operation.

fn reset(&mut self)

Reset the state.

fn status(&self) -> OptimizationStatus

Return the builder status.

fn properties(&self) -> OptimizationProperties

Return the builder properties.

fn len(&self) -> usize

The number of operation fused.

Provided Methods

fn is_empty(&self) -> bool

If no operations are fused.