Crate burn_fusion

Burn Fusion

This library is a part of the Burn project. It is a standalone crate that can be used to perform automatic operation fusion on backends that support it.

Modules

client

Client module exposing types to communicate with the fusion server.

graph

Graph module exposing all tensor operations that can be optimized.

Macros

binary_float_cmp_ops

binary_float_ops

binary_int_cmp_ops

binary_int_ops

scalar_float2int_ops

scalar_float_cmp_ops

scalar_float_ops

scalar_int_cmp_ops

scalar_int_ops

unary_float_ops

unary_int_ops

Structs

DeviceId

The device id.

Fusion

Enable dynamic operation fusion on a backend that implements fusion backend.

FusionTensor

Tensor primitive for the fusion backend for all kind.

HandleContainer

Keep all tensor handles in one place and ensure that all resources are used optimally.

OptimizationProperties

The properties of a builder.

TensorDescription

A tensor definition represents a snapshot of a tensor when it was used.

TensorId

The tensor unique identifier.

Enums

OptimizationStatus

The status of a builder.

TensorStatus

The status of the current tensor.

Traits

FusionBackend

Trait that allows an existing backend to specify graph optimizations using operation builder.

FusionDevice

The handle device trait allows to get an id for a backend device.

Optimization

The operation created from the builder.

OptimizationBuilder

The fusion operation abstraction allows implementations to fuse many tensor operations into one, improving the performance of the backend.

Type Aliases

Handle

Type alias for fusion backend handle.