Trait burn_fusion::client::FusionClient

pub trait FusionClient: Send + Sync + Clone {
    type FusionBackend: FusionBackend;

    // Required methods
    fn new(device: <Self::FusionBackend as FusionBackend>::FusionDevice) -> Self;
    fn register<O: Ops<Self::FusionBackend> + 'static>(
        &self,
        description: TensorOpsDescription,
        ops: O
    );
    fn drain_graph(&self);
    fn device(&self) -> &<Self::FusionBackend as FusionBackend>::FusionDevice;
    fn tensor_uninitialized(&self, shape: Vec<usize>) -> FusionTensor<Self>;
    fn register_tensor(
        &self,
        handle: Handle<Self::FusionBackend>,
        shape: Vec<usize>
    ) -> FusionTensor<Self>;
    fn read_tensor_float<const D: usize>(
        &self,
        tensor: TensorDescription
    ) -> Reader<Data<FloatElem<Self::FusionBackend>, D>>;
    fn read_tensor_int<const D: usize>(
        &self,
        tensor: TensorDescription
    ) -> Reader<Data<IntElem<Self::FusionBackend>, D>>;
    fn read_tensor_bool<const D: usize>(
        &self,
        tensor: TensorDescription
    ) -> Reader<Data<bool, D>>;
    fn change_client_float<const D: usize>(
        &self,
        tensor: TensorDescription,
        client: Self
    ) -> FusionTensor<Self>;
    fn change_client_int<const D: usize>(
        &self,
        tensor: TensorDescription,
        client: Self
    ) -> FusionTensor<Self>;
    fn change_client_bool<const D: usize>(
        &self,
        tensor: TensorDescription,
        client: Self
    ) -> FusionTensor<Self>;
    fn register_orphan(&self, id: &TensorId);

}

Define how to interact with the fusion server.

Required Associated Types

type FusionBackend: FusionBackend

The fusion backend associated type.

Required Methods

fn new(device: <Self::FusionBackend as FusionBackend>::FusionDevice) -> Self

Create a new client for the given fusion device.

fn register<O: Ops<Self::FusionBackend> + 'static>( &self, description: TensorOpsDescription, ops: O )

Register a new tensor operation description.

fn drain_graph(&self)

Register all lazy computation.

fn device(&self) -> &<Self::FusionBackend as FusionBackend>::FusionDevice

Get the current device used by all operations handled by this client.

fn tensor_uninitialized(&self, shape: Vec<usize>) -> FusionTensor<Self>

Create a new fusion tensor, but with no resources allocated to it.

fn register_tensor( &self, handle: Handle<Self::FusionBackend>, shape: Vec<usize> ) -> FusionTensor<Self>

Create a tensor with the given handle and shape.

fn read_tensor_float<const D: usize>( &self, tensor: TensorDescription ) -> Reader<Data<FloatElem<Self::FusionBackend>, D>>

Read the values contained by a float tensor.

fn read_tensor_int<const D: usize>( &self, tensor: TensorDescription ) -> Reader<Data<IntElem<Self::FusionBackend>, D>>

Read the values contained by an int tensor.

fn read_tensor_bool<const D: usize>( &self, tensor: TensorDescription ) -> Reader<Data<bool, D>>

Read the values contained by a bool tensor.

fn change_client_float<const D: usize>( &self, tensor: TensorDescription, client: Self ) -> FusionTensor<Self>

Change the client of the given float tensor.

fn change_client_int<const D: usize>( &self, tensor: TensorDescription, client: Self ) -> FusionTensor<Self>

Change the client of the given int tensor.

fn change_client_bool<const D: usize>( &self, tensor: TensorDescription, client: Self ) -> FusionTensor<Self>

Change the client of the given bool tensor.

fn register_orphan(&self, id: &TensorId)

Drop the tensor with the given tensor id.